backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
openai==1.3.0 